+++ src/PluginAuto/Mac/PluginWindowMacCarbonQD.h
@@ -30,27 +30,27 @@ namespace FB
         virtual ~PluginWindowMacCarbonQD();
 
         int16_t HandleEvent(EventRecord* evt);
-        void setWindowPosition(int x, int y, int w, int h);
-        void getWindowPosition(int &x, int &y, int &w, int &h) const;
-		int getWindowWidth() const { return m_width; }
-		int getWindowHeight() const { return m_height; }
+        void setWindowPosition(long x, long y, long w, long h);
+        void getWindowPosition(long &x, long &y, long &w, long &h) const;
+		long getWindowWidth() const { return m_width; }
+		long getWindowHeight() const { return m_height; }
 		FB::Rect getWindowPosition() const;
-        void setWindowClipping(int t, int l, int b, int r);
-        void getWindowClipping(int &t, int &l, int &b, int &r) const;
+        void setWindowClipping(long t, long l, long b, long r);
+        void getWindowClipping(long &t, long &l, long &b, long &r) const;
 		FB::Rect getWindowClipping() const;
         void InvalidateWindow() const {}
         CGrafPtr getPort() const { return m_port; }
     protected:
         CGrafPtr m_port;
-        int m_x;
-        int m_y;
-        int m_width;
-        int m_height;
+        long m_x;
+        long m_y;
+        long m_width;
+        long m_height;
         
-        int m_clipLeft;
-        int m_clipRight;
-        int m_clipTop;
-        int m_clipBottom;
+        long m_clipLeft;
+        long m_clipRight;
+        long m_clipTop;
+        long m_clipBottom;
     };
 };
 #endif // H_PLUGINWINDOWMAC_CARBON_QD
