+++ src/PluginAuto/Mac/PluginWindowMacCocoaCG.h
@@ -36,7 +36,7 @@ namespace FB {
             CGContextRef getContext() { return m_context; }
 
             void setWindowPosition(int32_t x, int32_t y, uint32_t width, uint32_t height);
-            void setWindowClipping(uint16_t top, uint16_t left, uint16_t bottom, uint16_t right);
+            void setWindowClipping(uint32_t top, uint32_t left, uint32_t bottom, uint32_t right);
             NPRect getWindowPosition();
             NPRect getWindowClipping();
