+++ src/PluginAuto/Mac/PluginWindowMacCocoaCG.mm
@@ -64,7 +64,7 @@ void PluginWindowMacCocoaCG::setWindowPosition(int32_t x, int32_t y, uint32_t wi
     m_height = height;
 }
 
-void PluginWindowMacCocoaCG::setWindowClipping(uint16_t top, uint16_t left, uint16_t bottom, uint16_t right) {
+void PluginWindowMacCocoaCG::setWindowClipping(uint32_t top, uint32_t left, uint32_t bottom, uint32_t right) {
     m_clipTop = top;
     m_clipLeft = left;
     m_clipBottom = bottom;
