+++ src/PluginAuto/Mac/PluginWindowMacQuickDraw.h
@@ -26,27 +26,27 @@ namespace FB {
     class PluginWindowMacQuickDraw : public PluginWindow
     {
     public:
-        PluginWindowMacQuickDraw(CGrafPtr port, int x, int y);
+        PluginWindowMacQuickDraw(CGrafPtr port, long x, long y);
         virtual ~PluginWindowMacQuickDraw();
 
         int16_t HandleEvent(EventRecord* evt);
-        void setWindowPosition(int x, int y, int w, int h);
-        void getWindowPosition(int &x, int &y, int &w, int &h);
-        void setWindowClipping(int t, int l, int b, int r);
-        void getWindowClipping(int &t, int &l, int &b, int &r);
+        void setWindowPosition(long x, long y, long w, long h);
+        void getWindowPosition(long &x, long &y, long &w, long &h);
+        void setWindowClipping(long t, long l, long b, long r);
+        void getWindowClipping(long &t, long &l, long &b, long &r);
         CGrafPtr getPort() { return m_port; }
         virtual void InvalidateWindow();
     protected:
         CGrafPtr m_port;
-        int m_x;
-        int m_y;
-        int m_width;
-        int m_height;
+        long m_x;
+        long m_y;
+        long m_width;
+        long m_height;
 
-        int m_clipLeft;
-        int m_clipRight;
-        int m_clipTop;
-        int m_clipBottom;
+        long m_clipLeft;
+        long m_clipRight;
+        long m_clipTop;
+        long m_clipBottom;
     };
 
 };
