+++ src/ScriptingCore/APITypes.h
@@ -342,10 +342,10 @@ namespace FB
     }
 
     struct Rect {
-        int16_t top;
-        int16_t left;
-        int16_t bottom;
-        int16_t right;
+        int32_t top;
+        int32_t left;
+        int32_t bottom;
+        int32_t right;
     };
 }
